clangビルド時のエラーを修正

+++ Source/Framework/Core/Public/Framework/Core/Network/TCPClient.h
@@ -45,7 +45,7 @@ namespace ob::core {
 	private:
 		IPAddress m_ip;
 		u16 m_port;
-		s32 m_socket;
+		u64 m_socket;
 	};
